refactor: 지역 정보 조회 컨트롤러 별도로 분리 (MIKKI-105)
